requirements.txt
python-dotenv==1.0.1
pandas
requests==2.31.0
requests-oauthlib==2.0.0
tzdata==2024.1
sqlalchemy
pyodbc
pymssql==2.2.7